[cleanup] moe: rename moelayer.py to moe_layer.py
