+++ fairscale/nn/moe/__init__.py
@@ -3,5 +3,5 @@
 # This source code is licensed under the BSD license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .moelayer import MOELayer
+from .moe_layer import MOELayer
 from .top2gate import Top2Gate
